+++ CHANGELOG.md
@@ -8,3 +8,4 @@
    - link accounts
 3. Adjust to client side verification
 4. Add twitter claim page
+5. Post a tweet to Claim, misc cleanup